+++ src/main/java/com/nekromant/telegram/service/MentoringSubscriptionRequestService.java
@@ -1,44 +1,27 @@
 package com.nekromant.telegram.service;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonParseException;
-import com.nekromant.telegram.MentoringReviewBot;
 import com.nekromant.telegram.commands.dto.ChequeDTO;
-import com.nekromant.telegram.commands.dto.LifePayResponseDTO;
-import com.nekromant.telegram.commands.feign.LifePayFeign;
 import com.nekromant.telegram.config.LifePayProperties;
 import com.nekromant.telegram.config.PriceProperties;
-import com.nekromant.telegram.contants.PayStatus;
 import com.nekromant.telegram.contants.ServiceType;
 import com.nekromant.telegram.model.MentoringSubscriptionRequest;
 import com.nekromant.telegram.model.PaymentDetails;
 import com.nekromant.telegram.repository.MentoringSubscriptionRequestRepository;
-import com.nekromant.telegram.repository.PaymentDetailsRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.dao.DataAccessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.Map;
 
 import static com.nekromant.telegram.contants.MessageContants.MENTORING_OFFER_DESCRIPTION;
-import static java.rmi.server.LogStream.log;
+import static com.nekromant.telegram.contants.MessageContants.RESPONSE_FOR_MENTORING_SUBSCRIPTION;
 
 @Slf4j
 @Service
-public class MentoringSubscriptionRequestService {
+public class MentoringSubscriptionRequestService extends ClientPaymentRequestServiceCommon implements ClientPaymentRequestService {
     @Autowired
     private MentoringSubscriptionRequestRepository mentoringSubscriptionRequestRepository;
-    @Autowired
-    private PaymentDetailsRepository paymentDetailsRepository;
-    @Autowired
-    private UserInfoService userInfoService;
-    @Autowired
-    private MentoringReviewBot mentoringReviewBot;
-    @Autowired
-    private LifePayFeign lifePayFeign;
     @Autowired
     private LifePayProperties lifePayProperties;
     @Autowired
@@ -57,30 +40,13 @@ public class MentoringSubscriptionRequestService {
                 mentoringData.get("PHONE").toString(),
                 lifePayProperties.getMethod());
 
-        try {
-            MentoringSubscriptionRequestService.log.info("Sending request to LifePay" + chequeDTO);
-            LifePayResponseDTO lifePayResponse = new Gson().fromJson(lifePayFeign.payCheque(chequeDTO).getBody(), LifePayResponseDTO.class);
+        return save(ServiceType.MENTORING, chequeDTO, mentoringSubscriptionRequest, mentoringSubscriptionRequestRepository, null);
+    }
 
-            MentoringSubscriptionRequestService.log.info("LifePay response: " + lifePayResponse);
-
-            PaymentDetails paymentDetails = PaymentDetails.builder()
-                    .number(lifePayResponse.getData().getNumber())
-                    .status(PayStatus.UNREDEEMED)
-                    .serviceType(ServiceType.MENTORING)
-                    .build();
-            paymentDetailsRepository.save(paymentDetails);
-            MentoringSubscriptionRequestService.log.info("Unredeemed payment created: " + paymentDetails);
-
-            mentoringSubscriptionRequest.setLifePayTransactionNumber(lifePayResponse.getData().getNumber());
-            mentoringSubscriptionRequestRepository.save(mentoringSubscriptionRequest);
-            MentoringSubscriptionRequestService.log.info("New mentoring subscription request created: " + mentoringSubscriptionRequest);
-            return ResponseEntity.ok(lifePayResponse.getData().getPaymentUrlWeb());
-        } catch (JsonParseException jsonParseException) {
-            log("Erorr while parsing Json: " + jsonParseException.getMessage());
-            return ResponseEntity.badRequest().build();
-        } catch (DataAccessException dataAccessException) {
-            log("Error while accessing database: " + dataAccessException.getMessage());
-            return ResponseEntity.internalServerError().build();
-        }
+    public void notifyMentor(PaymentDetails paymentDetails) {
+        String text = generateTextForMentoringBotNotification(paymentDetails,
+                RESPONSE_FOR_MENTORING_SUBSCRIPTION,
+                mentoringSubscriptionRequestRepository.findByLifePayTransactionNumber(paymentDetails.getNumber()).getTgName());
+        notifyMentor(paymentDetails, text);
     }
 }
