+++ src/main/java/com/nekromant/telegram/service/ResumeAnalysisRequestService.java
@@ -1,41 +1,40 @@
 package com.nekromant.telegram.service;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonParseException;
 import com.nekromant.telegram.commands.dto.ChequeDTO;
-import com.nekromant.telegram.commands.dto.LifePayResponseDTO;
 import com.nekromant.telegram.commands.feign.LifePayFeign;
+import com.nekromant.telegram.commands.feign.TelegramFeign;
 import com.nekromant.telegram.config.LifePayProperties;
-import com.nekromant.telegram.contants.PayStatus;
 import com.nekromant.telegram.contants.ServiceType;
 import com.nekromant.telegram.model.PaymentDetails;
-import com.nekromant.telegram.model.Promocode;
 import com.nekromant.telegram.model.ResumeAnalysisRequest;
 import com.nekromant.telegram.repository.PaymentDetailsRepository;
 import com.nekromant.telegram.repository.ResumeAnalysisRequestRepository;
+import feign.form.FormData;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataAccessException;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.util.Set;
+import javax.ws.rs.core.MediaType;
 
+import static com.nekromant.telegram.contants.MessageContants.RESPONSE_FOR_RESUME_PROJARKA;
 import static com.nekromant.telegram.contants.MessageContants.RESUME_OFFER_DESCRIPTION;
-import static java.rmi.server.LogStream.log;
 
 
 @Slf4j
 @Service
-public class ResumeAnalysisRequestService extends ClientPaymentRequestService {
+public class ResumeAnalysisRequestService extends ClientPaymentRequestServiceCommon implements ClientPaymentRequestService {
     @Autowired
     private ResumeAnalysisRequestRepository resumeAnalysisRequestRepository;
     @Autowired
-    private PaymentDetailsRepository paymentDetailsRepository;
-    @Autowired
     private PromocodeService promocodeService;
     @Autowired
-    private LifePayFeign lifePayFeign;
+    private UserInfoService userInfoService;
+    @Value("${owner.userName}")
+    private String ownerUserName;
+    @Autowired
+    private TelegramFeign telegramFeign;
     @Autowired
     private LifePayProperties lifePayProperties;
 
@@ -48,42 +47,25 @@ public class ResumeAnalysisRequestService extends ClientPaymentRequestService {
 
         ChequeDTO chequeDTO = new ChequeDTO(lifePayProperties.getLogin(),
                 lifePayProperties.getApikey(),
-                super.calculatePriceWithOptionalDiscount(CVPromocodeId, this.getClass().getSimpleName()),
+                calculatePriceWithOptionalDiscount(CVPromocodeId, this.getClass().getSimpleName()),
                 RESUME_OFFER_DESCRIPTION,
                 phone,
                 lifePayProperties.getMethod());
 
-        try {
-            ResumeAnalysisRequestService.log.info("Sending request to LifePay" + chequeDTO);
-            LifePayResponseDTO lifePayResponse = new Gson().fromJson(lifePayFeign.payCheque(chequeDTO).getBody(), LifePayResponseDTO.class);
-            ResumeAnalysisRequestService.log.info("LifePay response: " + lifePayResponse);
+        return save(ServiceType.RESUME, chequeDTO, resumeAnalysisRequest, resumeAnalysisRequestRepository, CVPromocodeId);
+    }
 
-            PaymentDetails paymentDetails = PaymentDetails.builder()
-                    .number(lifePayResponse.getData().getNumber())
-                    .status(PayStatus.UNREDEEMED)
-                    .serviceType(ServiceType.RESUME)
-                    .build();
-            paymentDetailsRepository.save(paymentDetails);
-            ResumeAnalysisRequestService.log.info("Unredeemed payment created: " + paymentDetails);
+    public void notifyMentor(PaymentDetails paymentDetails) {
+        promocodeService.incrementCounterUsed(paymentDetails);
 
-            resumeAnalysisRequest.setLifePayTransactionNumber(lifePayResponse.getData().getNumber());
-            resumeAnalysisRequestRepository.save(resumeAnalysisRequest);
-            ResumeAnalysisRequestService.log.info("New resume analysis request created: " + resumeAnalysisRequest);
+        String text = generateTextForMentoringBotNotification(paymentDetails,
+                RESPONSE_FOR_RESUME_PROJARKA,
+                resumeAnalysisRequestRepository.findByLifePayTransactionNumber(paymentDetails.getNumber()).getTgName());
+        notifyMentor(paymentDetails, text);
 
-            Promocode promocode = promocodeService.findById(CVPromocodeId);
-            if (promocode != null) {
-                Set<PaymentDetails> promocodePaymentDetailsSet = promocode.getPaymentDetailsSet();
-                promocodePaymentDetailsSet.add(paymentDetails);
-                promocode.setPaymentDetailsSet(promocodePaymentDetailsSet);
-                promocodeService.save(promocode);
-            }
-            return ResponseEntity.ok(lifePayResponse.getData().getPaymentUrlWeb());
-        } catch (JsonParseException jsonParseException) {
-            log("Erorr while parsing Json: " + jsonParseException.getMessage());
-            return ResponseEntity.badRequest().build();
-        } catch (DataAccessException dataAccessException) {
-            log("Error while accessing database: " + dataAccessException.getMessage());
-            return ResponseEntity.internalServerError().build();
-        }
+        String receiverId = userInfoService.getUserInfo(ownerUserName).getChatId().toString();
+        byte[] CV_bytes = resumeAnalysisRequestRepository.findByLifePayTransactionNumber(paymentDetails.getNumber()).getCVPdf();
+        FormData formData = new FormData(MediaType.MULTIPART_FORM_DATA, "document", CV_bytes);
+        telegramFeign.sendDocument(formData, receiverId);
     }
 }
